+++ SFMLTemplate/Constants.h
@@ -6,7 +6,7 @@ public:
 	Constants();
 	~Constants();
 	static void UpdateView(sf::View v);
-	static const unsigned short boardSize = 40;
+	static const int boardSize = 40;
 	static float tileSize;
 	static std::vector<std::string> attributes;
 	static sf::FloatRect fixedView;
